moved inf to constants
