+++ tidy3d/__init__.py
@@ -44,10 +44,7 @@ from .components import DATA_TYPE_MAP, ScalarFieldData, ScalarFieldTimeData
 from .components import ModeData, ModeAmpsData, ModeIndexData
 
 # constants imported as `C_0 = td.C_0` or `td.constants.C_0`
-from .constants import C_0, ETA_0, HBAR, EPSILON_0, MU_0, Q_e
-
-# types
-from .components import inf
+from .constants import C_0, ETA_0, HBAR, EPSILON_0, MU_0, Q_e, inf
 
 # plugins typically imported as `from tidy3d.plugins import DispersionFitter`
 from . import plugins
