+++ tidy3d/components/__init__.py
@@ -36,6 +36,3 @@ from .simulation import Simulation
 # data
 from .data import SimulationData, FieldData, FluxData, ModeData, FluxTimeData
 from .data import ScalarFieldData, ScalarFieldTimeData, ModeAmpsData, ModeIndexData, DATA_TYPE_MAP
-
-# types
-from .types import inf
